cleaning code

+++ app/api/products/route.ts
@@ -4,7 +4,7 @@ import Softwere from "@/models/softwere";
 import Product from "@/models/product";
 import connectMongoDB from "@/libs/mongodb";
 import { NextRequest, NextResponse } from "next/server";
-import { NextApiRequest, NextApiResponse } from "next";
+import { NextApiResponse } from "next";
 
 import {
   createAllProductsPipeline,
@@ -20,30 +20,17 @@ export async function GET(req: NextRequest, res: NextApiResponse) {
   const pageQuery = urlObject.searchParams.get("page");
   const collection = urlObject.searchParams.get("collection");
   const sort = urlObject.searchParams.get("sort");
-
   const page = (pageQuery !== null && parseInt(pageQuery)) || 1;
   const pageSize = 8;
-  let models;
-
-  switch (collection) {
-    case "allproducts":
-      models = [DJ, Vinyl, Softwere];
-      break;
-    case "vinyls":
-      models = [Vinyl];
-      break;
-    case "djequipments":
-      models = [DJ];
-      break;
-    case "softweres":
-      models = [Softwere];
-      break;
-    default:
-      models = [DJ, Vinyl, Softwere];
-      break;
-  }
+  const collectionMap: any = {
+    allproducts: [DJ, Vinyl, Softwere],
+    vinyls: [Vinyl],
+    djequipments: [DJ],
+    softweres: [Softwere],
+  };
+  const models =
+    collection !== null ? collectionMap[collection] : [DJ, Vinyl, Softwere];
   const newArrivalsPipeline = createNewArrivalsPipeline([DJ, Vinyl, Softwere]);
-
   const allProductsPipeline = createAllProductsPipeline(
     models,
     query,
@@ -52,7 +39,6 @@ export async function GET(req: NextRequest, res: NextApiResponse) {
     sort
   );
   const productsQuantityPipeline = createPagesNumberPipeline(models, query);
-
   try {
     const [newProducts, products, productsQuantity] = await Promise.all([
       Product.aggregate(newArrivalsPipeline),
@@ -60,15 +46,18 @@ export async function GET(req: NextRequest, res: NextApiResponse) {
       Product.aggregate(productsQuantityPipeline),
     ]);
 
-    return NextResponse.json({
-      products,
-      newProducts,
-      pages:
-        productsQuantity.length !== 0 &&
-        Math.ceil(productsQuantity[0].totalProducts / pageSize),
-    });
+    return NextResponse.json(
+      {
+        products,
+        newProducts,
+        pages:
+          productsQuantity.length !== 0 &&
+          Math.ceil(productsQuantity[0].totalProducts / pageSize),
+      },
+      { status: 201 }
+    );
   } catch (err) {
-    console.error("Error:", err);
+    console.log("Error message: Error with GET Products API");
   }
 }
 export async function POST(request: NextRequest, response: NextResponse) {
@@ -84,27 +73,15 @@ export async function POST(request: NextRequest, response: NextResponse) {
       inStock,
     } = await request.json();
 
-    let model;
-    let mongoCategory;
+    const categoryMap: any = {
+      dj: { model: DJ, mongoCategory: "djequipment" },
+      vinyls: { model: Vinyl, mongoCategory: "vinyl" },
+      softweres: { model: Softwere, mongoCategory: "softwere" },
+    };
 
-    switch (category) {
-      case "dj":
-        model = DJ;
-        mongoCategory = "djequipment";
-        break;
-      case "vinyls":
-        model = Vinyl;
-        mongoCategory = "vinyl";
-        break;
-      case "softweres":
-        model = Softwere;
-        mongoCategory = "softwere";
-        break;
-      default:
-        break;
-    }
+    const { model, mongoCategory } = categoryMap[category] || {};
 
-    await connectMongoDB(); // Connect to MongoDB
+    await connectMongoDB();
 
     if (model) {
       const createdProduct = await model.create({
@@ -118,17 +95,20 @@ export async function POST(request: NextRequest, response: NextResponse) {
         inStock,
       });
 
-      console.log("Product added successfully:", createdProduct._id); // Log the ID of the created product
+      console.log("Product added successfully");
       return NextResponse.json(
-        { id: createdProduct._id, message: "Product added successfully" },
+        {
+          id: createdProduct._id,
+          status: "success",
+        },
         { status: 201 }
       );
     } else {
-      console.error("Invalid category");
+      console.log("Error message: Invalid category on POST Products API");
       return NextResponse.json({ error: "Invalid category" }, { status: 400 });
     }
   } catch (error) {
-    console.error("Error adding product:", error);
+    console.log("Error message: Error with POST Products API");
     return NextResponse.json(
       { error: "Error adding product" },
       { status: 500 }
@@ -136,37 +116,27 @@ export async function POST(request: NextRequest, response: NextResponse) {
   }
 }
 export async function DELETE(request: NextRequest, response: NextResponse) {
-  console.log("here");
-
   const { id, category } = await request.json();
-  console.log(id, category);
-  let model;
 
-  switch (category) {
-    case "djequipment":
-      model = DJ;
-      break;
-    case "vinyl":
-      model = Vinyl;
-      break;
-    case "softwere":
-      model = Softwere;
-      break;
-    default:
-      break;
+  const categoryModelMap: any = {
+    djequipment: DJ,
+    vinyl: Vinyl,
+    softwere: Softwere,
+  };
+
+  const model = categoryModelMap[category];
+  if (!model) {
+    console.log("Error message: Invalid category on DELETE Products API");
+
+    return NextResponse.json({ message: "Invalid category" }, { status: 400 });
   }
-  console.log(typeof id);
-
-  // if (!model) {
-  //   return NextResponse.json({ message: "Invalid category" }, { status: 400 });
-  // }
 
   await connectMongoDB();
   if (model) {
     const product = await model.findByIdAndDelete(id);
-    // Handle product deletion or any other logic here
   } else {
-    // Handle the case where model is not defined
+    console.log("Error message: Product not found on DELETE Products API");
+
     return NextResponse.json({ message: "Product not found" }, { status: 404 });
   }
